+++ src/common.rs
@@ -24,19 +24,19 @@ pub struct SimpleDestructible {
 #[derive(Clone, Copy, Debug)]
 
 pub struct Countdown {
-    initial_value: usize,
-    current_value: usize,
+    initial_value: u32,
+    current_value: u32,
 }
 
 #[derive(Clone, Debug)]
 pub struct VariableCountdown {
-    range: Range<usize>,
-    current_value: usize,
+    range: Range<u32>,
+    current_value: u32,
     pristine: bool,
 }
 
 impl VariableCountdown {
-    pub fn new(range: Range<usize>) -> Self {
+    pub fn new(range: Range<u32>) -> Self {
         if range.start < 1 {
             panic!("Countdown range must have values above zero");
         }
@@ -53,7 +53,7 @@ impl VariableCountdown {
             self.current_value -= 1;
             return false;
         } else {
-            let initial_value = rng.usize(self.range.clone());
+            let initial_value = rng.u32(self.range.clone());
             self.current_value = initial_value;
             if self.pristine {
                 self.pristine = false;
@@ -66,7 +66,7 @@ impl VariableCountdown {
 }
 
 impl Countdown {
-    pub fn new(initial_value: usize) -> Self {
+    pub fn new(initial_value: u32) -> Self {
         Self {
             initial_value,
             current_value: initial_value,
