+++ src/resources.rs
@@ -21,7 +21,7 @@ pub enum ResourceKind {
 #[derive(Clone, Copy, Component, Debug)]
 pub struct ResourceChunk {
     pub kind: ResourceKind,
-    pub quantity: usize,
+    pub quantity: u32,
 }
 
 #[derive(Component)]
@@ -32,20 +32,20 @@ pub struct CarryingResources(pub Vec<ResourceChunk>);
 
 #[derive(Component)]
 pub struct ResourceCarrier {
-    pub max_volume: usize,
+    pub max_volume: u32,
 }
 
 pub struct BreaksIntoResourcesEvent(pub Entity);
 
 pub struct ResourcesPlugin;
 
-fn resource_volume_per_unit(resource_kind: ResourceKind) -> usize {
+fn resource_volume_per_unit(resource_kind: ResourceKind) -> u32 {
     match resource_kind {
         ResourceKind::Wood => 20,
     }
 }
 
-pub fn max_resource_units_from_chunk(chunk: &ResourceChunk, available_volume: usize) -> usize {
+pub fn max_resource_units_from_chunk(chunk: &ResourceChunk, available_volume: u32) -> u32 {
     (available_volume / resource_volume_per_unit(chunk.kind)).min(chunk.quantity)
 }
 
@@ -58,12 +58,12 @@ pub enum ResourceChunkAddingResult {
 pub fn add_resource_chunk(
     resource_chunks: &Vec<ResourceChunk>,
     added_resource_chunk: &ResourceChunk,
-    max_volume: usize,
+    max_volume: u32,
 ) -> ResourceChunkAddingResult {
     let current_volume = resource_chunks
         .iter()
         .map(|x| resource_volume_per_unit(x.kind) * x.quantity)
-        .sum::<usize>();
+        .sum::<u32>();
     let available_volume = max_volume - current_volume;
     let quantity_to_add = max_resource_units_from_chunk(added_resource_chunk, available_volume);
     if quantity_to_add == 0 {
